# FaxtimeRest/config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v1",
        "title": "TAIHOINST WEB API",
        "description": "A TAIHOINST WEB API based on Swagger and built using Node.js"
    },
    "paths": {
...
        "/v1/faxtime/sms/send/{id}": {
            "get": {
                "tags": [
                    "/V1/FAXTIME/SMS/SDK_SMS_SEND"
                ],
                "operationId": "FAXTIME_SDK_SMS_SEND_GETBYID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "description": "SMS 발송 데이터 조회",
                "parameters": [
                    { "$ref" : "#/parameters/agentParam" },
                    { "$ref" : "#/parameters/msgParam" }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/RESULTS"
                            }
                        }
                    }
                },
                "deprecated": false 
            },
            "put": {
                "operationId": "FAXTIME_SDK_SMS_SEND_PUTBYID",
                "tags": [
                  "/V1/FAXTIME/SMS/SDK_SMS_SEND"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    { "$ref" : "#/parameters/agentParam" },
                    { "$ref" : "#/parameters/msgParam" },
                    {
                        "name" : "datas",
                        "in" : "body",
                        "required" : true,
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/FAXTIME_SDK_SMS_DATA_POST"
                            }
                        }
                    }
                ],
                "description": "SMS 발송 데이터 수정",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/RESULTS"
                            }
                        }
                    }
                }
            },
            "delete": {
                "operationId": "FAXTIME_SDK_SMS_SEND_DELETEBYID",
                "tags": [
                  "/V1/FAXTIME/SMS/SDK_SMS_SEND"
                ],
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    { "$ref" : "#/parameters/agentParam" },
                    { "$ref" : "#/parameters/msgParam" }
                ],
                "description": "SMS 발송 데이터 삭제",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/RESULTS"
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/others/file/upload" : {
            "post" : {
                "tags": [
                    "/V1/OTHERS/FILE_UPLOAD"
                ],
                "operationId": "OTHERS_FILE_UPLOAD_POST",
                "consumes": [
                    "multipart/form-data",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json",
                    "text/html"
                ],
                "parameters": [
                    { "$ref" : "#/parameters/agentParam" },
                    {
                        "name": "objName",
                        "in": "query",
                        "description": "name...",
                        "required": true,
                        "type": "string",
                        "default" : "mmsfile1"
                    },
                    {
                        "name": "maxByte",
                        "in": "query",
                        "description": "byte...",
                        "required": false,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "subPath",
                        "in": "query",
                        "description": "path...",
                        "required": false,
                        "type": "string",
                        "default" : "mmsfile"
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "첨부파일",
                        "required": true,
                        "type": "file"
                    }
                ],
                "description": "파일업로드",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "items": {
                                "$ref": "#/definitions/RESULTS"
                            }
                        }
                    }
                }
            }
        }
    },
    "parameters": {
        "agentParam" : {
            "name": "Agent-Key",
            "in": "header",
            "description": "agent Identifier...",
            "required": true,
            "type": "string",
            "default" : "30748F97-429A-443B-BC66-9A32AD20B07F"
        },
...
        "msgParam": {
            "name": "id",
            "in": "path",
            "description": "message identifier...",
            "required": true,
            "type": "string"
        }
    },
    "securityDefinitions": {
        "apikeyQuery" : {
            "type": "apiKey",
            "in": "header",
            "name": "QUERY-API-KEY"
        },
        "userSecurity": {
            "type": "apiKey",
            "in": "header",
            "name": "USER-API-KEY"
        },
        "adminSecurity": {
            "type": "apiKey",
            "in": "header",
            "name": "ADMIN-API-KEY"
        },
        "agentSecurity": {
            "type": "apiKey",
            "in": "header",
            "name": "AGENT-API-KEY"
        },
        "mediaSecurity": {
            "type": "apiKey",
            "in": "query",
            "name": "MEDIA-API-KEY"
        }
    },
    "definitions": {
        "RESULTS": {
            "type": "object",
            "properties": {
                "status" : {
                    "type" : "string"
                },
                "description" : {
                    "type" : "string"
                },
                "datas" : {
                    "type" : "object"
                }
            }
        },
        "FAXTIME_SDK_SMS_DATA_POST" : {
            "type" : "object",
            "properties" : {
                "user" : {
                    "type" : "string"
                },
                "chkReserve" : {
                    "type" : "string"
                },
                "sTitle": {
                    "type" : "string"
                },
                "vmessage" : {
                    "type" : "string"
                },
                "revDttm" : {
                    "type" : "string"
                },
                "vsourcetel" : {
                    "type" : "string"
                },
                "vreceiver" : {
                    "type" : "string"
                },
                "vdestinationtel" : {
                    "type" : "string"
                }
            }
        },
...
    }
}